added logging
